# Conflicts:
#   src/main/kotlin/Nodes/Statement/WhileNode.kt

+++ src/main/kotlin/Nodes/Expressions/IntLitNode.kt
@@ -6,7 +6,7 @@ import main.kotlin.Utils.LitTypes
 import src.main.kotlin.Nodes.ExprNode
 
 
-class IntLitNode(val int_val : Int, override val ctx: BasicParser.IntLitContext) : ExprNode {
+class IntLitNode(val int_val : Long, override val ctx: BasicParser.IntLitContext) : ExprNode {
 
     override fun getType(): LitTypes {
         return LitTypes.IntWacc
